Make the kubernetes subdomain optional
Kubernetes subdomains are completely optional according to the
documentation on
https://kubernetes.io/docs/concepts/services-networking/dns-pod-service/.
Actually, the subdomain is required here and not having it will result
on the call returning `"null"` when you don't have a subdomain and hence
the hostname ends up being something like `VerneMQ@vernemq-0.null.default.svc.cluster.local`
that causes nodes not able to talk to each other. This fix makes sure to
handle both cases when there is a subdomain or the lack of it.