Improved hostname detection in DOCKER_VERNEMQ_SWARM mode (#373)
* Improved hostname detection in DOCKER_VERNEMQ_SWARM mode

Previous $(hostname -i) setup lead to a space separated hostnames like "VerneMQ@10.0.2.3 10.0.15.3" in a multi-network docker swarm configuration ( see #372 ). This lead to obvious crashing of the nodes. The suggested code picks eth0 as network for hostname detection in docker-swarm mode and therefore only supplies one IP.

* Updated hostname detection in docker-sawarm mode

Changed docker-swarm node hostname detection to only use first argument of "hostname -i" command, so multiple networks will be caught before ending up in the nodename.

* Updated hostname detection for Docker-Swarm mode

On second thought, using the network interface defined by the user as DOCKER_NET_INTERFACE or the one already acquired in the line 3 of this script might be better for hostname detection than to rely on the first output of "hostname -i" being the correct net interface.